+++ R/vcfR.R
@@ -9,8 +9,9 @@
 #' 
 #' \strong{File input and output} is facilitated with the functions \code{\link{read.vcfR}} and \code{\link{write.vcf}}.
 #' Input of vcf format data results in an S4 \code{\link{vcfR-class}} object.
-#' Objects of class vcfR can be manipulated with \code{\link{vcfR-method}}s and \code{\link{extract.gt}}.
-#' Contents of the vcfR object can be visualized with the \code{\link{plot.vcfR}} method.
+# ' Objects of class vcfR can be manipulated with \code{\link{vcfR-method}} and \code{\link{extract.gt}}.
+#' Objects of class vcfR can be manipulated with \link[vcfR:vcfR-method]{vcfR-method} and \code{\link{extract.gt}}.
+#' Contents of the vcfR object can be visualized with the \code{\link{plot}} method.
 #' More complex visualizations can be created using a series of functions.
 #' See \code{vignette(topic="sequence_coverage")} for an example.
 #' Once manipulations are complete the object may be written to a *.vcf.gz format file using \code{\link{write.vcf}} or exported to objects supported by other R packages with \code{\link{vcfR2genind}} or \code{\link{vcfR2loci}}.
@@ -41,7 +42,7 @@
 #' @docType package
 #' @name vcfR
 #' @rdname vcfR
-#' @useDynLib vcfR
+#' @useDynLib vcfR, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 #' @importFrom stats setNames
 #'