+++ tests/testthat.R
@@ -1,3 +1,4 @@
+
 library("testthat")
 #library("vcfR")
 #
@@ -14,12 +15,15 @@ test_file("testthat/test_2_chromR.R")
 test_file("testthat/test_3_extract_gt.R")
 #
 test_file("testthat/test_4_vcfR2DNAbin.R")
+#test_file("testthat/test_io.R")
 #
-test_file("testthat/test_io.R")
+test_file("testthat/test_proc_chromR.R")
 #
 test_file("testthat/test_conversion.R")
 #
 test_file("testthat/test_vcfR_methods.R")
+#
+test_file("testthat/test_chromR_method.R")
 #
 test_file("testthat/test_addID.R")
 #
@@ -28,26 +32,30 @@ test_file("testthat/test_maf.R")
 test_file("testthat/test_genotype_matrix_functions.R")
 #
 test_file("testthat/test_vcfRtidy.R")
-#
-test_file("testthat/test_heatmapbp.R")
+
 #
 test_file("testthat/test_ad_frequency.R")
-#
-test_file("testthat/test_chromR_method.R")
-#
-test_file("testthat/test_get.R")
+
+
 #
 test_file("testthat/test_freq_peak.R")
+#
+test_file("testthat/test_is_het.R")
+#
+test_file("testthat/test_get.R")
+
+test_file("testthat/test_genetic_diff.R")
+
+#test_file("testthat/test_heatmapbp.R")
 
 
 # This test will write the plots to a file on the filesystem.
 # This should cause CRAN to complain.
 # Comment out this test prior to CRAN submission.
 #
-#
-test_file("testthat/test_chromo_plot.R")
-#
-test_file("testthat/test_drplot.R")
+#test_file("testthat/test_chromo_plot.R")
+#test_file("testthat/test_drplot.R")
+
 
 
 ##### ##### ##### ##### #####
@@ -55,7 +63,6 @@ test_file("testthat/test_drplot.R")
 
 #test_file("testthat/test_rank_variants.R")
 #test_file("testthat/test_var_window.R")
-#test_file("testthat/test_windowing.R")