+++ tests/testthat/test_2_chromR.R
@@ -236,7 +236,7 @@ test_that("chromR2vcfR works",{
   test <- chromR2vcfR(chrom, use.mask = TRUE)
 
   sum(chrom@var.info$mask)
-nrow(test)
+  nrow(test)
 })
 
 ##### ##### ##### ##### #####