+++ tests/testthat/test_get.R
@@ -3,7 +3,11 @@ library('testthat')
 
 context("getFIX accessors")
 
+
+
+#
 data("vcfR_test")
+#
 chrom <- create.chromR(vcfR_test, verbose = FALSE)
 
 test_that("getFIX returns a matrix", {